chore :: 이미지 파일 구조 변경
