+++ components/layout/header/style.ts
@@ -1,11 +1,9 @@
-import React from 'react';
 import styled from '@emotion/styled';
-import { THEME, Z_INDEX } from '../../styles/theme';
-import { Button } from '../ui/Button';
-import Image from 'next/image';
+import { THEME, Z_INDEX } from '../../../styles/theme';
+
 
 // 헤더 전체 래퍼
-const HeaderWrapper = styled.header`
+export const HeaderWrapper = styled.header`
   width: 100%;
   height: 60px;
   display: flex;
@@ -20,7 +18,7 @@ const HeaderWrapper = styled.header`
 `;
 
 
-const HeaderContent = styled.div`
+export const HeaderContent = styled.div`
   width: 100%;
   max-width: 1920px;
   height: 100%;
@@ -31,7 +29,7 @@ const HeaderContent = styled.div`
 `;
 
 // 로고
-const Logo = styled.div`
+export const Logo = styled.div`
   display: flex;
   align-items: center;
   font-family: ${THEME.typography.fontFamily[0]};
@@ -43,14 +41,14 @@ const Logo = styled.div`
 `;
 
 // 메뉴 래퍼
-const MenuWrapper = styled.nav`
+export const MenuWrapper = styled.nav`
   display: flex;
   align-items: center;
   gap: 2.5rem;
 `;
 
 // 메뉴 텍스트
-const MenuItem = styled.a`
+export const MenuItem = styled.a`
   font-family: ${THEME.typography.fontFamily[0]};
   font-size: 1rem;
   font-weight: 400;
@@ -62,24 +60,3 @@ const MenuItem = styled.a`
     color: ${THEME.colors.Primary.Normal};
   }
 `;
-
-const Header: React.FC = () => {
-  return (
-    <HeaderWrapper>
-      <HeaderContent>
-        <Logo>
-          <Image src="/logo.svg" alt="Calyx Logo" width={42} height={45} priority />
-        </Logo>
-        <MenuWrapper>
-          <MenuItem href="#">기능</MenuItem>
-          <MenuItem href="#">대시보드</MenuItem>
-          <MenuItem href="#">사용법</MenuItem>
-          <MenuItem href="#">후기</MenuItem>
-          <Button variant="primary" size="md" className="pill">시작하기</Button>
-        </MenuWrapper>
-      </HeaderContent>
-    </HeaderWrapper>
-  );
-};
-
-export default Header;
